Moving jit code into qirlib
