+++ pyqir-jit/Cargo.toml
@@ -12,10 +12,6 @@ readme = "README.md"
 [dependencies]
 qirlib = { path = "../qirlib" }
 inkwell = { git = "https://github.com/TheDan64/inkwell", branch = "master", default-features = false, features = ["llvm11-0", "target-x86"] }
-lazy_static = "1.4.0"
-mut_static = "5.0.0"
-log = "0.4.14"
-bitvec = "1.0.0"
 
 [dependencies.pyo3]
 version = "0.15.1"
