+++ qirlib/Cargo.toml
@@ -14,6 +14,7 @@ rand="0.8.4"
 base64 = "0.13.0"
 lazy_static = "1.4.0"
 mut_static = "5.0.0"
+bitvec = "1.0.0"
 
 [dev-dependencies]
 serial_test = "0.5.1"
