+++ qirlib/resources/tests/custom_entry_point_name.bc
