+++ qirlib/resources/tests/entry_point_types.bc
