+++ qirlib/resources/tests/multiple_entry_points.bc
