+++ qirlib/src/lib.rs
@@ -4,6 +4,7 @@
 #![warn(clippy::all, clippy::pedantic)]
 
 pub mod codegen;
+pub mod evaluation;
 pub mod generation;
 pub mod module;
 pub mod passes;
